Sketch out the basic structure.
This is currently capable of running hello world and nothing beyond that. Even
then, it cheats in a few ways. However, this is much more about getting a
minimalish skeleton of a VM implemented that we can start expanding and
correcting.

The key component is the memory design: we have a reference counted GC that
exposes trait objects as single machine word big pointers and that only requires
a single malloc per object. Getting to this point has involved a number of false
starts: the eventual design is simple and, I hope, relatively non-clunky to use.
In essence, the `Gc` type in `gc.rs` is roughly similar to `Rc` but: you can
turn a `Gc` into a pointer and back into a cloned-and-incremented `Gc` in one
go; and using `GcBox::alloc_blank` you can allocate a `Gc` of arbitrary size
(which is very handy when you're creating trait objects). The `ThinObj` type in
`vm/objects.rs` then stores away vtable pointers so that we end up with thin
pointers. The most annoying hack is the `GcLayout` trait, and even that's not
*too* bad. More to the point, the design should hopefully allow us to move
fairly easily to a proper GC in the future.

In terms of the more yksom-specific things, the key is probably in
`vm/objects.rs`. The `Val` struct hides whether it is storing a boxed object or
an unboxed integer (or so on). Boxed objects implement the `Obj` trait. The key
function is then `Val::gc_obj` which returns a boxed object (boxing as
necessary). Structuring things in this way makes it relatively easy to use
things with the borrow checker. `Obj`s have a method `as_any` which casts them
to `Any`: this allows you to take an `Obj`, convert it to an `Any`, and then use
`downcast_ref` to safely dynamically get at the particular struct (`Class`,
`String_` and so on) inside it.

Currently we only implement a miniscule (and not correct -- println should be on
System, not on String) subset of SOM's standard library in `lib/SOM`. This is
enough to run hello world:

```
Hello = (
    run = ('Hello, World from SOM' println )
)
```

One nice touch relative to other SOMs is that we get proper parsing error
messages through lrpar (and a less broken parser than SOM).

Anyway, there's still a lot to do here, and I'm sure that some parts of this
design will prove to be bone-headed. Which, of course, remains to be seen...