Fully free memory from Gc/GcBox/ThinObj.
For a little while I wasn't really sure if this was possible, because Box::drop
in the standard library has this wonderful comment:

  // FIXME: Do nothing, drop is currently performed by compiler.

Fortunately ptr::drop_in_place does exactly what we want by some magic process
that I don't understand. [If you want to be bamboozled, look at the source for
core::ptr::real_drop_in_place.]

With this, we drop every bit of memory we allocate.