Allow object references to be converted back into ThinCells/Gcs.
Without this we can't write VM functions that return &self as a Val, because
&self will be a reference to a concrete struct, and we can't turn that into a
Val, because Val::from_obj takes a T, not an &T. At its simplest, this means we
couldn't write the following identity function:

  struct S;

  impl Obj for S {}

  impl S {
      fn identity(&self) -> Val {
          // what goes here?
      }
  }

This commit adds a way to turn a &T back into a ThinObj/GcBox/Gc. It allows us
to express the identity function as follows:

  impl S {
      fn identity(&self) -> Val {
          Val::restore(self)
      }
  }

This is clearly dangerous, because it assumes that the memory pointed to by
&self is part of a ThinObj/GcBox combo. If that isn't true, all sorts of bad
things will happen. Fortunately it's relatively easy to obey this constraint by
always passing new instances immediately to Val::from_obj.

[I spent considerable time working out if I could enforce this constraint in the
type system. My conclusion is that I probably could, but at the cost of making
the code hugely unergonomic.]