Provide a nicer way of casting to a concrete Obj type.
There is some inherent ickiness here, because given a `Val`, you always have to
call `gc_obj` in order to get a `ThinObj`. From Rust's perspective, the
resulting `ThinObj` is a temporary value (which can be quite confusing when you
start borrowing things).

Perhaps surprisingly, this gives us an opportunity: we can't put casting on
`Val` because if it's unboxed, we have to create a boxed version, which creates
a temporary which we can't then borrow from. However, by putting `cast` on
`ThinObj` we sidestep the problem entirely: we get a (relatively) nice way of
casting (including generating a `VMError` which records what type we expected,
and what we actually got) while avoiding introducing any new borrow checking
problems.