Use a custom, faster, casting mechanism.
This is a cunning hack based on the fact that vtable pointers are a proxy for a
type identifier. In other words, if two distinct objects have the same vtable
pointer, they are instances of the same type; if their vtable pointers are
different, they are instances of different types. We can thus use vtable
pointers as a proxy for type equality.

This commit moves the `cast` function over to do this, avoiding the slower `Any`
type (which has to create more than one trait object). I created a simple test
program which loops a SOM program that uses some casting. Before this commit
(on, I admit, my laptop) the timings are as follows:

```
===> multitime results
1: -q target/release/yksom --cp lib/SOM SOM/Examples/Hello.som
            Mean        Std.Dev.    Min         Median      Max
real        0.372       0.006       0.367       0.371       0.395
user        0.313       0.025       0.260       0.310       0.370
sys         0.063       0.025       0.020       0.060       0.110
```

After:

```
===> multitime results
1: -q target/release/yksom --cp lib/SOM SOM/Examples/Hello.som
            Mean        Std.Dev.    Min         Median      Max
real        0.348       0.006       0.340       0.346       0.364
user        0.282       0.022       0.240       0.280       0.340
sys         0.070       0.022       0.030       0.070       0.110
```

That's a ~6-7% performance in a program which is spending a lot of its time
doing non-casting things. In other words, the previous mechanism was really
very slow!

Admittedly this technique relies on the assumption that the same type always has
the same vtable pointer. This isn't guaranteed anywhere, and there's no way I
can think of to test for this property. But it seems fairly unlikely that it
would be violated, for the simple reason that it would worsen performance and
increase code bloat.