Banish TypeId.
TypeId seemed like the solution to several of our problems in that it allows us
a nice way to compare a static type and the run-time type of a value.
Unfortunately, `TypeId::of` can only be called on types with a 'static lifetime
(i.e. types which do not contain non-static references in them). This is too
severe a restriction for us.

We thus introduce our own simple type hierarchy `ObjType`: for any `Obj` we can
ask for either its static type (via `StaticObjType::static_objtype()`) or its
dynamic type (via `Obj::dyn_objtype()`). Notice that we need to create a new
trait for this (`StaticObjType`) because the sole function it requires prevents
trait objects being made (since it doesn't have a `&self` paramater).

Overall, then, this commit gets us the same functionality as `TypeId` while
freeing us from the restriction of having to make `Obj`s have 'static lifetimes.