Allow GCBOX Vals to be cast directly to references without going via tobj.
`Val::tobj` can box unboxed items but, because it does so, creates temporary
values that mean one can't store a reference to the underlying object.
`Val::gcbox_cast` is a much more limited function in that it can't box unboxed
items, but it doesn't create a temporary value, so one can then get a reference
to the underlying object. In general, though, `Val::tobj` remains the preferred
way of getting at objects.

This then allows us to store a class's superclass as a reference, rather than a
Val. It's only a tiny performance difference, but since it's in a key part of
the VM, every little counts.