Sort out the defining of names.
For reasons now lost in the past, previously we had two very similar rules for
defining a class's instance fields and a method's variables (confusingly called
"temps" -- it probably made sense at the time).

This commit simplifies the grammar, removing the duplicate name definition
stuff, and calling it "NameDefs". This has a small knock on effect on the
compiler.