Add a simple test runner for language-level tests.
We really need to be able to test what the compiler's doing, particularly with
errors, and it's annoying to write tests for such things in the normal Rust
framework. This simple test runner -- loosely based on that which comes with
Rust -- runs all `*.som` files in the `lang_tests` directory and checks their
output.

There are still some things which are a bit awkward to test in this framework
(e.g. any test with `"` inside a SOM comment), but this seems like a reasonable
start that we can build upon.

In the future, once we can run the main SOM test harness, we can get rid of some
of these tests (e.g. things like "hello world"), but until then this serves as a
useful sanity check for features that we otherwise can't test.