Support user calling / defining of multi-arg functions.
Although this isn't optimised yet, I've embedded some strong assumptions in the
compiler / VM such that we can do so later: by guaranteeing that the first
variable is self and the next n variables are the function's arguments (in
reverse) order, we give ourselves the possibility of doing a memcpy between a
(user) function's frame and the next frame rather than having to execute lots of
separate VM instructions. At the moment this isn't taken advantage of (indeed,
we are very inefficient, creating a temporary `Vec`!).