Implement non-local returns.
In SOM, a block can return from the outer function, so a function like:

  f = (
      [ ^'a' ] value.
      ^ 'b'.
  )

returns 'a', not 'b'. Note that blocks can be nested arbitrarily deep before a
non-local return occurs. This is a bit tricky, since blocks are also functions
of a sort, so we have to handle this carefully, with a little bit of compiler
support (so that we know how deeply nested a block is for a non-local return)
and VM support.

This raises one not-completely-obvious question, notably what happens when a
block escapes? Consider:

  escaped1 = (
      f = ( ^[ ^ 'a' ] )

      run = (
          self f value
      )
  )

We only call the block in line 2 once `f` has finished. We thus need to detect
that the block has escaped its original context (we use our closure pointers to
detect this) and throw an error. At the moment we simply panic: we can deal with
it more elegantly in the future. The various lang_tests/[block|escaped]*.som
tests are consistent with java-som.

This commit does require a bit of refactoring, notably making the `Frame` class
use inner mutation (which it should probably have always used). I'm also not
convinced that this commit is the most elegant way to handle non-local returns
(in particular, it might be more elegant to have a loop to handle layered
function calls, thus removing the need for the VMError::Return hack), but it at
least minimally works and moves us forwards.