Don't use cargo run in lang tests.
`cargo run` has a noticeable overhead of about 0.1s vs. running a binary
directly. Normally this isn't a problem worth bothering about, but since we're
starting to grow an ever larger number of language tests, this starts to add up.
On my main development branch (which has 22 language tests) on a 4 core machine,
this cuts the testing time down from 3s to 2s. That might not sound like much,
but we're only going to grow more and more language tests over time...