Val's cannot contain references.
We previously, unintentionally, undermined Rust's type system: you could have an
object with lifetime 'a, move it into a `Val`, and the lifetime disappeared.
This commit minimally fixes the only instance of this (in `Class`) and prevents
anyone doing this again by having `Val::from_obj` force a `'static` lifetime on
anything from which a `Val` is created.