Simplify casting.
This is a deceptively simple commit and I can't believe it didn't occur to me
earlier. In essence, we add a new trait `NotUnboxable` which if implemented for
a SOM object means "this struct can never be unboxed". In our case, we implement
this trait for everything except `Int`s. If you want to downcast a `Val` to a
given `NotUnboxable` object type you can then change:

    let str_tobj = rcv.tobj(self)?;
    let str_: &String_ = downcast(&str_tobj)?;

into:

    let str_: &String_ = rcv.gcbox_downcast(vm)?;

Not only is this much simpler to write and read, but a) we can now statically
detect when you're trying to do an illegal `gcbox_downcast` (e.g. you get a
compile-time error if you try to downcast to an `Int`) b) we remove a check and
branch from the performance critical `gcbox_downcas` function. So this seems
like a win all round!