Move Val into its own module.
objects.rs was starting to feel too big and complicated: breaking Val out is the
obvious place to start. This commit is fairly minimal in that it simply breaks
Val (and a few associated bits out), and fixes the resulting errors and
warnings, but doesn't make things quite as tidy as one might like.