Use a custom, compact representation of Result<Val, Box<VMError>>.
The problem we had before is that `VMError` is, naturally, a fairly large enum:
on a 64 bit machine it is currently 32 bytes. Before this commit, we have lots
of functions with a return type of `Result<Val, VMError>` -- which is 40 bytes
big! This is going to cause things to continually spill to the stack, which is
not good for performance.

This commit does two things (in one go, because it made more sense that way): it
changes `Result<Val, VMError>` into `Result<Val, Box<VMError>>` and then makes a
new type `ValResult` which is a custom representation of that type. `ValResult`
is a machine word big, so in essence this commit turns the 40 bytes of
`Result<Val, VMError>` into the 8 bytes of `ValResult`. The strong working
assumption is that `Box<VMError>` is an unusual case and that we don't mind
spilling it to the heap when it does happen. That is, mostly, a reasonable
assumption *except* that currently SOM closures (not SOM functions) use/abuse
VMError to percolate their return value up the call stack. That's something to
fix for another day.

The underlying technique is pretty simple: we have spare tag bits in `Val`, so
we can use one of those to differentiate between a `Val` and a `ValResult`. This
means that we get compile-time checking that we're not using a `Val` where we
wanted a `ValResult`, but the two both share the same underlying representation
(a machine word).