Add support for looping.
The "restart" primitive is unusual in structure: it affects the *current* frame.
That require us to do a bit of code shuffling. I don't particularly like the
impact this has on the code, but the alternative is to continuously write "pc"
into/out of a frame which doesn't seem very good either. In other words, this is
something that we might want to revisit at some point.