Implement block parameters.
The way SOM handles these is a bit cack-handed: a block with no parameters is of
type Block, a block with 1 parameters is of type Block2, and a block with 2
parameters of type Block3.

The problem then becomes where the best place to encode this in the VM is e.g.
we could duplicate code (and make another dent in the icache), or we could
introduce a branch (and confuse the branch predictor with a semi-random branch)
or etc.

In the end, I chose what I hope is the last worst option: we introduce a branch
at the point of block creation and a reference to the specific block class in
the block itself. That means that creating a block is slightly more expensive
than before, but calling it is as cheap as before. Given that blocks are used as
while bodies, this seems like a reasonable trade-off.