Implement arbitrary sized integers.
This allows arithmetic operations to operate on numbers that are as large as you
have RAM for (i.e. you're not restricted to working on integers in machine
words). We use the `num_bigint` library to represent such arbitrary sized
integers.

Arbitrary sized integers are not an efficient way of representing small numbers,
so we now have three ways of representing integers:

  * Tagged fixed-sized integers (i.e. an isize minus 3 bits)
  * Boxed fixed-sized integers (boxed isizes)
  * Boxed arbitrary-sized integers (boxed BigInts)

For any given number we always try to store it as the most efficient
representation (i.e. if "x" is a BigInt and you do "x-x" you'll get back a
tagged integer representing 0). This is completely transparent to the user: SOM
simply exposes an "integer" type and doesn't differentiate between fixed-sized
and arbitrary-sized integers.