Rework number operations.
Our previous way of performing numeric operations was fragile, largely because it
could auto-cast one numeric type (e.g. an `Int`) to another (e.g. an `ArbInt`).
It then became very difficult to get all the possible conversions correct.

This commit reworks that. We no longer auto-cast between numeric types: users
need to do that themselves.

As part of that, it is now natural to handle tagged integers in the `Val` which
is more efficient. A simple `while` loop which counts down from 100000 to 0 is
now 5% faster because it doesn't have to continually box values. [It's still
slow in an overall sense because we haven't optimised SOM's closures/stack in
any way yet.]

So this commit can feel a bit weird: it makes things more complex in order to
make them simpler and, as a side bonus, also makes things a bit faster.