Duplicate everything in the Obj trait in Val.
The aim here is that "normal" VM code (i.e. outside `Val`) almost never calls
`tobj`. So what was e.g.:

  rtry!(obj.tobj(vm)).get_class(vm)

now becomes:

  obj.get_class(vm)

This means that we're more efficient when handling tagged objects (we don't need
to box an integer merely to find out its class), and we then hide the need to
box things from the user in many more circumstances.

In practise, we'd almost got to this point anyway, and only the get_class and
to_strval methods needed to be added.