Add reference equality.
This is roughly equivalent to the `is` operator in Python in that it checks
normal objects for pointer equality, but numbers are checked for value equality.
Note that SOM (or, at least, java-som) doesn't define/implement the latter cases
properly yet (see https://github.com/SOM-st/SOM/issues/23) so this commit makes
us diverge slightly.