Minimally break apart objects.rs.
This file was starting to get unwieldy (at just under 1000 lines) and it will
keep growing bigger over time, so now seems as good a time as any to split it
up. This commit does so as minimally as possible: there's still tidying up to
do, but this commit aims only to move code, and remove all warnings.

The main thing to note is that `integers.rs` contains both `Int` and `ArbInt`
since they're somewhat tied together (SOM code can't tell the two apart).