Primitives only make sense within an existing function frame.
My original assumption with SOM primitives was that they were "genuine" function
calls (creating a function frame and so on). I'm now reasonably sure that I was
wrong to think that: all the primitives I've implemented so far expect to
operate within the current function frame. Doing so will be much more efficient
than creating a new frame.

This commit is an intermediate step on the way. It removes the need to call
primitives in their own function frames, so exec_primitive now pushes results
directly onto the stack. However, it does lead to some code duplication and
general ugliness (in particular, exec_primitive now returns a Result type, which
doesn't compose very well with ValResult). Since it passes all the tests, it's a
useful commit in its own right.