Remove ValResult.
This was introduced because I thought we would need a ternary return type to
deal with SOM's closure returns. However, the recent reorganisation of the core
interpreter loop has removed that need (the ternary aspect now being confined
purely to the few functions related to message sends). We now only need a
binary return type which means we're better off simply using the normal `Result`
type (which has the pleasant bonus of allowing us to use the `?` operator).

This commit thus removes the custom `ValResult` type and replaces it with
`Result<Val, Box<VMError>>`. Most of this commit was performed by `srep`, with a
small number of fix-ups by hand.

There appears to be a tiny performance benefit from this commit, but it's small
enough that it's not really worth worrying about (my guess is that in a couple
of places rustc can statically remove `Result` types which it couldn't do with
`ValResult`).