Reading from an unassigned SOM variable should return 'nil'.
Previously we threw an error in such cases. While, arguably, that might be
better language design, it doesn't match SOM's semantics. So this commit fixes
that.

The implementation is not particularly efficient in a reference counted GC (we
increment nil's reference count only to decrement it on the first assignment to
a var), but it's probably not worth optimising this at this stage.