Implement inline caching.
Inline caching remembers the last receiver class and method used at a given call
site: if the same receiver class is used again, the method is used directly
without performing the (relatively slow) method lookup sequence. If a different
receiver class is used from the cache, or if the cache is empty, the cache is
updated with the current receiver class and method. This is a powerful
optimisation (though it can be defeated by pathological code which continually
cycles between different receiver classes, as indeed the `ic1.som` test does),
giving a performance boost of well over 20% to our simple while loop.

The "bit_eq" function is both extremely safe and extremely dangerous: safe in
the sense that it can't go wrong, dangerous in the sense that it might not give
people the properties they're expecting. The unusual name, and the documentation
string, will hopefully make people aware that it is not a function intended to
be used in many places.