Move to a single global vector of instructions and inline caches.
This commit is a two-in-one, since both parts (instructions and inline caches)
were tightly coupled (for not-very-interesting reasons surrounding `Copy` /
`Clone`).

Previously we had a per-class vector of instructions: this commit moves to a
single global vector of instructions. As part of that we also have to have a
single global vector of sends (so that `Instr` can become `Clone`).