Make method execution only require method pointers.
Previously, in order to execute a method you had to carry around its class,
since the method's program counter was relative to the class -- in other words
one would use a `(Class, Method)` pair to execute a method. Our recent changes
mean that everything we need to execute a method is now stored globally in the
core VM, so we no longer need to pass around a method's class in order to
execute the method.

This gives a roughly 8% speedup on our simple while benchmark, probably largely
because it reduces the amount of reference counting increments/decrements we
perform.