Move the UnsafeCell from SOMStack to VM.
This is how I should have done things in the first place, but I'd forgotten one
of my guiding principles, which is that `VM` should (roughly) model what a
multi-threaded interpreter should look like. In that model, some sort of lock
would be need to held before `SOMStack` can be accessed, but after that
`SOMStack` should be fairly normal.

This commit thus makes VM store `SOMStack` in an `UnsafeCell`, which means we
can then remove the inner `UnsafeCell` from `SOMStack` itself.