Build even when rustfmt can't be installed.
Sometimes rust nightly and/or rustfmt are so broken that the latter won't
install under any method we try. Rather than borking completely, it's better to
allow the PR to be merged even without the fmt check. This is less than ideal,
but I can't see a better way of doing things.