Block values should have self set to the object they were created in.
The `value` primitive is special, and before this commit we didn't implement it
correctly. The new test instance_var_in_block.som caused this error:

  thread 'main' panicked at 'called `Result::unwrap()` on an `Err` value:
  TypeError { expected: Inst, got: Block }', src/lib/vm/core.rs:353:39

The problem is that when executing a block via `value` we previously set `rcv` to
the `Block` itself.

This commit is a simple fix, effectively adding the active `self` to `Block` at
creation time so that we can fish it back out later. We can probably handle this
in a better way with closures later if we're so minded.