Try and get a handle on symbols / globals.
This is a fairly confusing part of SOM and we did not previously handle it
properly. Assuming "unknown" is a variable not used in the program, then this
code:

  unknown println.

is illegal but:

  global: #unknown

is legal (returning 'nil') and:

  unknown := ...

is illegal (complaining about "no field").

The key is to separate out symbols from globals as they're not really related:
in order to retain our previous optimisations (which, in general, make global
lookups an integer lookup), this does mean that `vm.globals` can contain
"illegal" values, which is icky, but I'm not sure I currently see a better way
of doing things.

There is more we could do here, in particular we currently don't really handle
the relationship between strings and symbols quite properly, though it's
surprisingly hard to notice the seams.