Implement class loading.
In SOM, classes are loaded when a Global hasn't had a value put into it.

Currently our implementation has one O(n) part to it: it's not super important
at this point, and I haven't quite worked out what the right design is, as we're
in danger of having too many caches and reverse caches.