Make the VM struct require mutable references.
This commit minimally (though that's a relative term -- this is a large patch,
which is a combination of srep, Vim search and replace, Vim macros, and hand
edits) moves yksom to the point where it can look more like normal Rust code
with less use of `UnsafeCell`. The key to doing that is to have the `VM` struct
require `&mut self`. That makes a lot of other code slightly more verbose, and I
quickly ran out of the ability to find meaningful variable names for all the
temporaries that one has to create to satisfy the borrow-checker -- `v` is now a
very common variable name.

I no longer know for sure why I didn't do this from the start, because even the
initial commit has VM taking `&self`. Partly, I think, it was because I couldn't
get some code past the-then borrow checker (I think I started pre-NLL Rust),
partly because I didn't think it was a good model of a multi-threaded VM.

That latter idea now seems a bit silly. SOM is fundamentally single-threaded,
but it's relatively easy to imagine the following model:

```
  struct VM {
    // stuff behind UnsafeCell
  }

  impl VM {
    fn f(&self) { ... }
  }

  struct VMThread<'a> {
    vm: &'a VM,
    // stuff not behind UnsafeCell
  }

  impl VMThread {
    fn f(&mut self) { ... }
  }
```

where most VM stuff would be defined on `VMThread`, including the core
interpreter loop, the thread-local VM stack etc. If one needed something from
the "global" VM struct then that would require mutexes / atomics to synchronise
and `UnsafeCell`s somewhere (probably inside `VM`'s functions). In other words,
I think a multi-threaded VM could mostly be written in normal-Rust style, with
only the small amount of global stuff having to use `UnsafeCell` or similar.