Remove the UnsafeCell around sends and reverse_sends.
Interestingly, this is the first example where we have to do something special
to please the borrow-checker: we have to put an `Rc` (or `Gc`, or similar)
around the `String` in `self.sends` so that we can lookup a method (which
requires `&mut VM`). We know that the `Rc` is redundant, because we never remove
`String`s from `self.sends`, but there's no way to statically assure the borrow
checker of that.