Remove mutability where it is not needed.
This is a bit of an experiment and might turn out to be a bad idea, in the sense
that if we *do* need to use a mutable VM object in any of these methods, we'll
then have to percolate that upwards. However, since rewriting for mutability
seems to mostly be fairly easy, this means that we can make some code a little
easier to write.