Remove `is_str` from `String_` and embed its class directly.
SOM has `String` and `Symbol` types where the latter is a subclass of the
former. This is a bit awkward to model in Rust. Previously `String_`s had an
`is_str` boolean (true for `String`, false for `Symbol`), which meant that we
ended up branching on it quite often.

This PR changes that so the creator of a string has to specify if they're
creating a SOM `String` (`String_::new_str`) or a `Symbol` (`String_::new_sym`),
allowing us to embed the appropriate class in the `String_` struct itself. This
doesn't increase the size of the struct (since the `is_str` bool was rounded up
to a machine word anyway), and it removes all the subsequent branches.

This does however make bootstrapping the VM a bit messier.