Introduce InstanceTypeError.
We need to deal with two kinds of errors, at least from the VM's internal
perspective: BuiltinTypeError (I expected you to pass me an object that's really
a Class struct) and InstanceTypeError (I expected you to pass me a Symbol but
you gave me a String).

The end user doesn't necessarily need to be aware of the difference between
these two things, because to some extent it's an artefact of the way we
implement things in Rust. Notably in SOM `Symbol` is a subclass of `String`, but
Rust has no direct way to model this; in 55985cea we embedded the class for
String/Symbols in the `String_` struct. Thus sometimes in the VM we want an
instance of the Rust `String_` struct that's an instance of the SOM `Symbol`
class (or SOM `String` class): these are `InstanceTypeError`s.

As an example of this, fix the long-ignored test that throws an
`InstanceTypeError` when one tries to lookup a global using a string instead of
a symbol.