super sends.
In SOM, "super" is mostly an alias for "self" *except* when you get a message
send such as "super x" which means "call the method 'x' in my superclass (or, if
not found, recursively in its superclasses)". This allows us to make another
part of Symbol.som identical to the SOM standard library.