from_[iu]size can't fail so don't have them return a Result type.
This changes from_usize from:

    pub fn from_usize(vm: &mut VM, i: usize) -> Result<Val, Box<VMError>>

to:

    pub fn from_usize(vm: &mut VM, i: usize) -> Val

and from_isize similarly. When we didn't have arbitrary sized integers, these
methods could fail, but it's been a long time since that was possible, and at
least one point in the code said "from_isize can't fail"!

This commit changes the signatures appropriately, which removes a lot of
`unwrap`s (mostly, admittedly, in tests) at the expenses of requiring quite a
lot of `Ok`s to be added around -- this seems like a fair swap to me, since
branches (in `unwrap`s) are generally more expensive than `Ok`s.