Fix incorrect asserts.
These asserts lacked tests that caused them to fire, and hence the fact that
they were incorrect had slipped through the net. Oops. Ironically, the asserts
were wrong because I overthought them: the whole point of this assert is to
check that we only make ArbInts for numbers that can't fit in an isize, so each
assert is identical.