Use SOM's initialize: method to start programs.
SOM starts programs by calling the initialize: method in System.som -- we now do
this in main.rs. This means that we now start SOM programs more-or-less in the
same way as other SOMs.

A few things needed tweaking to enable this, mostly making sure that we fix up
class references to method and class names after bootstrapping. The other major
addition is that SOM appears to treat String equality and reference equality as
being the same (or Java SOM might intern strings?), so we now do the same.