Use the Array trait to introduce a MethodsArray struct.
For the time being, this is functionally indistinguishable from a NormalArray:
they both respond to the same SOM methods. However, we will soon use the
MethodsArray struct to solve the problem of cache invalidation.

The way this works is as follows. Given a `Val`, one can now (via a `ThinObj`)
get a reference to the object as an `Array` trait object via the `to_array`
function in the `Obj` trait. [This slightly annoying dance is necessary because
one cannot otherwise coerce one trait object to another in Rust. By introducing
this method, and having the structs `impl` it, we use normal (and safe!) Rust
mechanisms to achieve the effect we want.]