When a class's methods are mutated, flush all inline caches.
yksom's use of inline caches (like most/all other SOMs) is beneficial to
performance, but leads to a correctness problem when a program mutates the
array resulting from the "methods" function call. mutate_methods.som in Java
SOM, for example, prints #b #b as if the method had not been mutated. Before
this commit, yksom also printed the same.

This commit makes use of MethodsArray to flush every inline cache in the VM
whenever a method array is mutated. By flushing every inline cache, we easily
solve the problem of changes to a superclass's methods being reflected in
subclasses. This is a sledgehammer to crack a nut, but it is a complex nut, and
rarely encountered. More elegant solutions are possible.