Call escapedBlock when a block escapes.
This means that we need to record in the frame the current block being executed
(if one exists). That's mildly annoying in that each frame takes an additional
two machine words (`Option<Val>` is not kind to us!), although it does mean that
we can drop the slightly-weird `is_method` bool in `Frame::new`.