Always terminate the VM by throwing an Exit exception.
This allows us to hack in backtraces even when SOM tries (not very effectively)
to tell the user what went wrong. For example, SOM catches "method not
understood" errors and prints a message (to stdout...) but only tells you what
the method name was. This commit means that, provided SOM calls `system exit: 1`
(or some other non-zero code) even in such cases you get a backtrace such as:

  ERROR: Method f not found in class perform_unknown
  Traceback (most recent call at bottom):
    File SOM/Smalltalk/System.som, line 46, column 8:
            (application respondsTo: #run:)
                ifTrue:  [ application run: arguments ]
                ifFalse: [ application run ]
    File SOM/Smalltalk/Boolean.som, line 31, column 8:
            self ifFalse: [ ^falseBlock value ].
    File SOM/Smalltalk/False.som, line 33, column 24:
        ifFalse: block = ( ^block value )
    File SOM/Smalltalk/Boolean.som, line 31, column 25:
            self ifFalse: [ ^falseBlock value ].
    File SOM/Smalltalk/System.som, line 48, column 23:
                ifFalse: [ application run ]
    File /home/ltratt/yksom/lang_tests/perform_unknown.som, line 9, column 9:
            (self perform: #f) println.
    File SOM/Smalltalk/Object.som, line 73, column 8:
            self error:
                ('Method ' + selector + ' not found in class ' + self class name)
    File SOM/Smalltalk/Object.som, line 64, column 64:
        error: string = ( '' println. ('ERROR: ' + string) println. system exit: 1 )
  exit(1).

The "reason" given is the code passed to `system exit:`. This is less good than
the backtrace support we had before for such messages, but it's better than
nothing.